old-tests/submission/Infraware/Forms/contents/Forms/Not Use/input_defaultchecked.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>input_defaultChecked</h3>
	</p>
	
	<li>The checkbox control be checked when the document was loaded</li>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  id="input_form">
	<p><input type='checkbox' id='input_checkbox'></p>
	</form>

	<script>
		
		var input_checkbox = document.getElementById("input_checkbox");
		input_checkbox.defaultChecked = true;

		if (typeof(input_checkbox.defaultChecked) == "boolean") {
			if (input_checkbox.defaultChecked != true) {
				test(function() {
					assert_true(false, "defaultChecked attribute is not correct.");
				});
			}
		} else {
			test(function() {
				assert_unreached("defaultChecked attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
